[Decomp][Unit] Cohesion split: extract spell bonus/mod methods to UnitSpellBonus.cpp
